Refactor code to support more sample code
