+++ gradle.properties
@@ -17,3 +17,6 @@ POM_DEVELOPER_NAME=Stas Parshin
 
 BINTRAY_REPO=maven
 BINTRAY_LICENSE=Apache-2.0
+
+android.useAndroidX=true
+android.enableJetifier=true
