+++ sample/build.gradle
@@ -24,4 +24,5 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'com.google.android.gms:play-services-maps:17.0.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+    implementation "androidx.constraintlayout:constraintlayout:2.0.0-rc1"
 }
