+++ sample/src/main/java/com/github/pengrad/mapscaleview/sample/MapboxActivity.kt
@@ -1,0 +1,11 @@
+package com.github.pengrad.mapscaleview.sample
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MapboxActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_mapbox)
+    }
+}
