+++ sample/src/main/res/layout/activity_main.xml
@@ -1,136 +1,29 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.github.pengrad.mapscaleview.sample.MainActivity">
+    tools:context=".MainActivity">
 
-    <fragment
-        android:id="@+id/mapFragment"
-        class="com.google.android.gms.maps.SupportMapFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <LinearLayout
+    <androidx.appcompat.widget.AppCompatButton
+        android:id="@+id/googlemaps"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
-        tools:ignore="HardcodedText">
+        android:text="@string/googlemaps"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/mapbox"/>
 
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeColor"
-            android:text="color" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeTextSize"
-            android:text="text size" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeStrokeWidth"
-            android:text="stroke width" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeMiles"
-            android:text="miles" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeMeters"
-            android:text="meters" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeSize"
-            android:text="size" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeOutline"
-            android:text="outline" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeDirection"
-            android:text="direction" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeFont"
-            android:text="font" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeTileSize"
-            android:text="tile size"/>
-
-    </LinearLayout>
-
-    <LinearLayout
+    <androidx.appcompat.widget.AppCompatButton
+        android:id="@+id/mapbox"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="top|end">
+        android:text="@string/mapbox"
+        app:layout_constraintTop_toBottomOf="@id/googlemaps"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"/>
 
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewBgFixed"
-            android:layout_width="100dp"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            android:background="#3333" />
-
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewBg"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            android:background="#3333"
-            app:scale_expandRtl="true" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="8dp"
-        android:orientation="vertical">
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewMiles"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:scale_miles="true"
-            app:scale_outline="true" />
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewRtl"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:scale_color="#dc5533"
-            app:scale_expandRtl="true" />
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:scale_color="#339cdc"
-            app:scale_textSize="20sp" />
-
-    </LinearLayout>
-
-</FrameLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
